not really sure tbh

+++ include/math_helpers.h
@@ -4,7 +4,7 @@
 #include <QPoint>
 
 QPointF midpoint(QPointF *A, QPointF *B);
-double distance(QPointF *A, QPointF *B);
+float distance(QPointF *A, QPointF *B);
 bool areCollinear(QPointF *A, QPointF *B, QPointF *C);
 
 #endif
